requirements.txt
Django==4.2.2
djangorestframework==3.14.0
psycopg2==2.9.6
drf-spectacular==0.26.3
pandas==2.0.2
openpyxl==3.1.2
